\name{EstHer-package}
\alias{EstHer-package}
\alias{EstHer}
\docType{package}
\title{
Estimation of Heritability in High Dimensional Sparse Linear Mixed Models using Variable Selection
}
\description{
Our method is a variable selection method whicg can be used to select active components in sparse linear mixed models in order to estimate the heritability. The selection allows us to reduce the size of the data sets which improves the accuracy of the estimations. Our package also provides a confidence interval for the estimated heritability.
}
\details{
\tabular{ll}{
Package: \tab EstHer\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-16\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Anna Bonnet and Celine Levy-Leduc

Maintainer: Anna Bonnet <anna.bonnet@agroparistech.fr>
~~ The author and/or maintainer of the package ~~
}
\references{
A.Bonnet, C.Levy-Leduc, E.Gassiat, R.Toro, T.Bourgeron. "Improving heritability estimation by a variable selection approach in sparse high dimensional linear mixed models", arXiv:1507.06245, 2015.
}
\keyword{Heritability estimation}
\keyword{sparsity}
\keyword{linear mixed models}
\keyword{variable selection}

