\name{X}
\alias{X}
\docType{data}
\title{
Matrix of fixed effects
}
\description{
Matrix of fixed effects of size n x d where n=200 and d=2
}
\usage{data("X")}
\format{
   The format is:
 num [1:200, 1:2] 1 1 1 1 1 1 1 1 1 1 ...
}
\examples{
data(X)
## maybe str(X) ; plot(X) ...
}
\keyword{datasets}
