\name{Y}
\alias{Y}
\docType{data}
\title{
Vector of observations
}
\description{
Vector of observations of size n=200.
}
\usage{data("Y")}
\format{
  The format is:
 num [1:200, 1] 6.196076 7.225162 9.662264 3.134030 5.404111 ...
}
\examples{
data(Y)
## maybe str(Y) ; plot(Y) ...
}
\keyword{datasets}
